Combining all BitSet types into a single class
It was becoming too complicated to have multiple BitSet types. While I liked having strong guarentees of the size, it became unwieldy. Consider if a Bus or a Register wants to hold multiple types. Generics can help, but this quickly falls apart if we want to create new BitSets of the expected type on the fly.